// src/app/core/interfaces/notification-preferences.ts
export interface NotificationPreferences {
  Topic: boolean;
  Discussion: boolean;
  DiscussionComment: boolean;
  TopicDiscussion: boolean;
  CommentVote: boolean;
  TopicReport: boolean;
  TopicVoteList: boolean;
  TopicEvent: boolean;
  Ideation: boolean;
  Idea: boolean;
  IdeaVote: boolean;
  TopicIdeation: boolean;
  IdeaComment: boolean;
  IdeaReport: boolean;
  CommentReport: boolean;
}